Resolve bug when loading CSV to ffproble instance.

+++ FFBitrateViewer.ApplicationAvalonia/Services/ffprobe/FFProbeAppClient.Models.cs
@@ -98,7 +98,7 @@ namespace FFBitrateViewer.ApplicationAvalonia.Services.ffprobe
         double? DTSTime,
 
         [property: JsonPropertyName("duration")]
-        int? Duration,
+        double? Duration,
 
         [property: JsonPropertyName("duration_time")]
         double? DurationTime,
